refactor: extract page selection buttons into a separate component
